[AV2-1007]
