+++ projects/v3/src/app/components/components.module.ts
@@ -15,13 +15,15 @@ import { LockTeamAssessmentPopUpComponent } from './lock-team-assessment-pop-up/
 import { ActivityCompletePopUpComponent } from './activity-complete-pop-up/activity-complete-pop-up.component';
 import { FastFeedbackComponent } from './fast-feedback/fast-feedback.component';
 import { ReviewRatingComponent } from './review-rating/review-rating.component';
+import { QuestionsModule } from '../questions/questions.module';
 
 @NgModule({
   imports: [
     IonicModule,
     CommonModule,
     FormsModule,
-    // FilestackModule
+    QuestionsModule,
+    // FilestackModule,
   ],
   declarations: [
     DescriptionComponent,
