+++ projects/v3/src/app/pages/home/home.module.ts
@@ -6,7 +6,6 @@ import { HomePageRoutingModule } from './home-routing.module';
 
 import { HomePage } from './home.page';
 import { ComponentsModule } from '../../components/components.module';
-import { PersonalisedHeaderComponent } from '@v3/app/components/personalised-header/personalised-header.component';
 
 @NgModule({
   imports: [
@@ -14,6 +13,6 @@ import { PersonalisedHeaderComponent } from '@v3/app/components/personalised-hea
     ComponentsModule,
     HomePageRoutingModule
   ],
-  declarations: [HomePage, PersonalisedHeaderComponent]
+  declarations: [HomePage]
 })
 export class HomePageModule {}
